secrets/crypto/alt/requirements.txt
coolname
cryptography
PyNaCl
